message

+++ 04-process/05-my-kill.js
@@ -5,6 +5,11 @@
 var pid = process.argv[2],
     sgn = process.argv[3];
 
+if(process.argv.length !==4||isNaN(Number(pid))){
+  console.error('命令行参数错误！');
+  process.exit(1);
+}
+
 
 process.kill(sgn,pid);
